requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dnspython==2.2.1
Flask==2.2.2
gevent==22.10.2
greenlet==2.0.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
pycparser==2.21
pymongo==4.2.0
python-dotenv==0.20.0
rel==0.4.7
requests==2.28.1
simple-websocket-server==0.4.2
urllib3==1.26.12
websocket-client==1.4.2
Werkzeug==2.2.2
zope.event==4.5.0
zope.interface==5.5.2